pyproject.toml
[project]
name = "ml-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "autogluon-tabular>=1.1.1",
    "fastai>=2.7.18",
    "mlflow>=2.17.2",
    "neuralprophet>=0.4.2",
    "optuna>=4.5.0",
    "pandas>=2.0.3",
    "pycaret>=2.2.2",
    "torch==2.0.1",
    "torchaudio==2.0.2",
    "torchvision==0.15.2",
    "transformers>=4.46.3",
]
